+++ fe/src/page/tour/ReviewList.jsx
@@ -1,5 +1,8 @@
 import React from "react";
 import ReviewItem from "./ReviewItem.jsx";
+import PropTypes from "prop-types";
+
+ReviewItem.prototype = { review: PropTypes.any };
 
 function ReviewList({ tourId, reviewList }) {
   return (
